+++ src/extension.ts
@@ -1,36 +1,22 @@
 'use strict';
 import MarkDownDOM from 'markdown-dom';
-import { ExtensionContext, window, workspace, Uri, TreeDataProvider, TreeItem, Event, TextDocument, EventEmitter, TreeItemCollapsibleState, ThemeIcon, commands, Selection, Range, RelativePattern } from 'vscode';
+import { ExtensionContext, window, workspace, Uri, TreeDataProvider, TreeItem, TextDocument, EventEmitter, TreeItemCollapsibleState, ThemeIcon, commands, Selection, Range, RelativePattern, FileSystemWatcher } from 'vscode';
 import * as path from 'path';
 
-type File = { type: 'file'; path: string; headlessTodos: Todo[]; heads: Head[]; };
-type Head = { type: 'head'; text: string; line: number; file: File; todos: Todo[]; };
-type Todo = { type: 'todo'; text: string; isChecked: boolean; line: number; file: File; indent: string; };
+const FileType: 'file' = 'file';
+type File = { type: typeof FileType; path: string; headlessTodos: Todo[]; heads: Head[]; };
+const HeadType: 'head' = 'head';
+type Head = { type: typeof HeadType; text: string; line: number; file: File; todos: Todo[]; };
+const TodoType: 'todo' = 'todo';
+type Todo = { type: typeof TodoType; text: string; isChecked: boolean; line: number; file: File; indent: string; };
 type Item = File | Head | Todo;
 
 export async function activate(context: ExtensionContext): Promise<void> {
-    context.subscriptions.push(commands.registerCommand('markdown-todo.openFile', (file: File) => {
-        window.showTextDocument(Uri.file(file.path), { preview: true });
-    }));
-
-    context.subscriptions.push(commands.registerCommand('markdown-todo.focusHead', async (head: Head) => {
-        const textEditor = await window.showTextDocument(Uri.file(head.file.path), { preview: true });
-        const range = textEditor.document.lineAt(head.line).range;
-        textEditor.selection = new Selection(range.end, range.start);
-        textEditor.revealRange(range);
-    }));
-
     context.subscriptions.push(commands.registerCommand('markdown-todo.focusTodo', async (todo: Todo) => {
         const textEditor = await window.showTextDocument(Uri.file(todo.file.path), { preview: true });
         const range = textEditor.document.lineAt(todo.line).range;
         textEditor.selection = new Selection(range.end, range.start);
         textEditor.revealRange(range);
-    }));
-
-    context.subscriptions.push(commands.registerCommand('markdown-todo.refreshFile', async (file: File) => {
-        //const textDocument = await workspace.openTextDocument(Uri.file(file.path));
-        // TODO: Fix the bug where some files do not update removed todos todoTreeDataProvider.refresh(textDocument);
-        await todoTreeDataProvider.refreshAll();
     }));
 
     context.subscriptions.push(commands.registerCommand('markdown-todo.removeTodo', async (todo: Todo) => {
@@ -40,6 +26,8 @@ export async function activate(context: ExtensionContext): Promise<void> {
         await textEditor.edit(editBuilder => {
             editBuilder.replace(range, '');
         });
+
+        await textEditor.document.save();
     }));
 
     context.subscriptions.push(commands.registerCommand('markdown-todo.toggleTodo', async (todo: Todo) => {
@@ -55,40 +43,53 @@ export async function activate(context: ExtensionContext): Promise<void> {
             editBuilder.replace(checkRange, todo.isChecked ? ' ' : 'x');
         });
 
-        // Update view without saving the file
-        todo.isChecked = !todo.isChecked;
-        // TODO: Fix the bug where some files do not update removed todos todoTreeDataProvider.refresh(textEditor.document);
-        await todoTreeDataProvider.refreshAll();
+        await textEditor.document.save();
     }));
 
     const todoTreeDataProvider = new TodoTreeDataProvider();
     context.subscriptions.push(todoTreeDataProvider);
     context.subscriptions.push(window.createTreeView('to-do', { treeDataProvider: todoTreeDataProvider }));
-
-    workspace.onDidSaveTextDocument(async textDocument => {
-        // Ignore irrelevant documents that will not affect the to-do items
-        if (textDocument.uri.scheme !== 'file' || textDocument.languageId !== 'markdown') {
-            return;
-        }
-
-        // TODO: Fix the bug where some files do not update removed todos todoTreeDataProvider.refresh(textDocument);
-        await todoTreeDataProvider.refreshAll();
-    });
-
-    await todoTreeDataProvider.refreshAll();
 }
 
 class TodoTreeDataProvider implements TreeDataProvider<Item> {
-    private cache: File[] = [];
+    private cache: Item[] = [];
     private _onDidChangeTreeData: EventEmitter<Item | undefined> = new EventEmitter<Item | undefined>();
-    readonly onDidChangeTreeData?: Event<Item | null | undefined> | undefined = this._onDidChangeTreeData.event;
+    public readonly onDidChangeTreeData = this._onDidChangeTreeData.event;
 
-    getTreeItem(element: Item) {
+    private readonly watcher: FileSystemWatcher;
+
+    constructor() {
+        this.watcher = workspace.createFileSystemWatcher('**/*.md');
+
+        this.watcher.onDidChange(async uri => {
+            this.refresh(await workspace.openTextDocument(uri));
+        });
+
+        this.watcher.onDidCreate(async uri => {
+            this.refresh(await workspace.openTextDocument(uri));
+        });
+
+        this.watcher.onDidDelete(uri => {
+            const index = this.cache.findIndex(item => item.type === 'file' && item.path === uri.fsPath);
+            this.cache.splice(index, 1);
+            this._onDidChangeTreeData.fire();
+        });
+
+        this.index();
+    }
+
+    public getTreeItem(element: Item) {
         switch (element.type) {
             case 'file': {
-                const count = element.headlessTodos.length + element.heads.reduce((count, head) => count + head.todos.length, 0);
-                const item = new TreeItem(`${path.parse(element.path).name} (${count})`, TreeItemCollapsibleState.Expanded);
-                item.command = { title: 'Open file', command: 'markdown-todo.openFile', arguments: [element] };
+                const headlessCounts = this.count(element.headlessTodos);
+                const headfulCounts = element.heads.reduce((counts, head) => {
+                    const { checked, unchecked } = this.count(head.todos);
+                    return { checked: counts.checked + checked, unchecked: counts.unchecked + unchecked };
+                }, { checked: 0, unchecked: 0 });
+                const checked = headlessCounts.checked + headfulCounts.checked;
+                const unchecked = headlessCounts.unchecked + headfulCounts.unchecked;
+                const total = checked + unchecked;
+                const item = new TreeItem(`${path.parse(element.path).name} (${checked} done, ${unchecked} to do, ${total} total)`, TreeItemCollapsibleState.Expanded);
                 item.contextValue = 'file';
                 item.iconPath = ThemeIcon.Folder;
                 item.id = element.path;
@@ -96,8 +97,9 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
                 return item;
             }
             case 'head': {
-                const item = new TreeItem(`${element.text} (${element.todos.length})`, TreeItemCollapsibleState.Expanded);
-                item.command = { title: 'Focus header', command: 'markdown-todo.focusHead', arguments: [element] };
+                const { checked, unchecked } = this.count(element.todos);
+                const total = checked + unchecked;
+                const item = new TreeItem(`${element.text} (${checked} done, ${unchecked} to do, ${total} total)`, TreeItemCollapsibleState.Expanded);
                 item.contextValue = 'head';
                 item.iconPath = ThemeIcon.Folder;
                 item.id = element.file.path + ':' + element.line;
@@ -120,7 +122,7 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
         }
     }
 
-    getChildren(element?: Item | undefined) {
+    public getChildren(element?: Item | undefined) {
         if (element === undefined) {
             return this.cache.map(file => file);
         }
@@ -136,11 +138,9 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
         // Todos do not have children.
     }
 
-    async refreshAll() {
-        this.cache = [];
-
+    private async index() {
         // TODO: https://github.com/Microsoft/vscode/issues/48674
-        const excludes = await workspace.getConfiguration('search').get('exclude')!;
+        const excludes = await workspace.getConfiguration('search', null).get('exclude')!;
         const globs = Object.keys(excludes).map(exclude => new RelativePattern(workspace.workspaceFolders![0], exclude));
         const occurences: { [fsPath: string]: number; } = {};
         for (const glob of globs) {
@@ -152,16 +152,25 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
 
         // Accept only files not excluded in any of the globs
         const files = Object.keys(occurences).filter(fsPath => occurences[fsPath] === globs.length);
-        console.log(files);
         for (const file of files) {
             const textDocument = await workspace.openTextDocument(Uri.file(file));
             this.refresh(textDocument);
         }
+
+        this._onDidChangeTreeData.fire();
     }
 
-    refresh(textDocument: TextDocument) {
-        const headlessTodos: { text: string; isChecked: boolean; line: number; indent: string; }[] = [];
-        const heads: { text: string; line: number; todos: { text: string; isChecked: boolean; line: number; indent: string; }[] }[] = [];
+    private refresh(textDocument: TextDocument) {
+        const path = textDocument.uri.fsPath;
+        const index = this.cache.findIndex(item => item.type === 'file' && item.path === path);
+        let file = this.cache[index] as File | undefined;
+        if (file !== undefined) {
+            file.headlessTodos = [];
+            file.heads = [];
+        } else {
+            file = { type: FileType, path, headlessTodos: [], heads: [] };
+        }
+
         for (let index = 0; index < textDocument.lineCount; index++) {
             const line = textDocument.lineAt(index);
             const dom = MarkDownDOM.parse(line.text.trim());
@@ -183,7 +192,8 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
                             text = text.substring(trash.length);
                         }
 
-                        heads.push({ text, line: line.lineNumber, todos: [] });
+                        const head: Head = { type: HeadType, text, line: line.lineNumber, todos: [], file };
+                        file.heads.push(head);
                         break;
                     }
                     case 'unordered-list': {
@@ -192,11 +202,11 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
                             if (item.type === 'checkbox') {
                                 // TODO: Figure out why item.indent is always zero
                                 const indent = (/^\s+/.exec(line.text) || [''])[0];
-                                const todo = { text: item.text.trim(), isChecked: item.check !== null, line: line.lineNumber, indent };
-                                if (heads.length === 0) {
-                                    headlessTodos.push(todo);
+                                const todo: Todo = { type: TodoType, text: item.text.trim(), isChecked: item.check !== null, line: line.lineNumber, indent, file };
+                                if (file.heads.length === 0) {
+                                    file.headlessTodos.push(todo);
                                 } else {
-                                    heads[heads.length - 1].todos.push(todo);
+                                    file.heads[file.heads.length - 1].todos.push(todo);
                                 }
                             }
                         } else {
@@ -209,28 +219,41 @@ class TodoTreeDataProvider implements TreeDataProvider<Item> {
             }
         }
 
-        if (headlessTodos.length > 0 || heads.reduce((count, head) => count + head.todos.length, 0) > 0) {
-            const path = textDocument.uri.fsPath;
-            let file = this.cache.find(item => item.type === 'file' && item.path === path);
-            if (file === undefined) {
-                file = { type: 'file', path, headlessTodos: [], heads: [] };
-                this.cache.push(file);
+        file.heads = file.heads.filter(head => head.todos.length > 0);
+
+        if (index !== -1) {
+            // Remove file after ites last to-do item has been deleted
+            if (file.headlessTodos.length === 0 && file.heads.length === 0) {
+                this.cache.splice(index, 1);
+                this._onDidChangeTreeData.fire();
+            } else {
+                this._onDidChangeTreeData.fire(file);
             }
-
-            file.headlessTodos = headlessTodos.map(todo => ({ ...todo, file: file!, type: 'todo' as 'todo' }));
-            file.heads = heads.filter(head => head.todos.length > 0).map(head => ({
-                type: 'head' as 'head',
-                ...head,
-                file: file!,
-                todos: head.todos.map(todo => ({ ...todo, file: file!, type: 'todo' as 'todo' }))
-            }));
+        } else {
+            // Do not include empty files
+            if (file.headlessTodos.length !== 0 || file.heads.length !== 0) {
+                this.cache.push(file);
+                // Refresh the tree to find the new file
+                this._onDidChangeTreeData.fire();
+            }
         }
-
-        this._onDidChangeTreeData.fire();
     }
 
-    dispose() {
+    private count(todos: Todo[]) {
+        return todos.reduce((counts, todo) => {
+            if (todo.isChecked) {
+                counts.checked++;
+            } else {
+                counts.unchecked++;
+            }
+
+            return counts;
+        }, { checked: 0, unchecked: 0 });
+    }
+
+    public dispose() {
         delete this.cache;
         this._onDidChangeTreeData.dispose();
+        this.watcher.dispose();
     }
 }